refactor: extract storage assertions
